+++ src/control_plane/project_infrastructure/plan_meilisearch_project_resources.rs
@@ -67,7 +67,7 @@ pub(crate) fn plan_meilisearch_project_resources(
     Ok(PreparedProjectService::new(
         project_id.to_owned(),
         service_id.to_owned(),
-        credential,
+        Some(credential),
         environment,
         container_environment,
         None,