+++ src/control_plane/project_infrastructure/plan_soketi_project_resources.rs
@@ -81,7 +81,7 @@ pub(crate) fn plan_soketi_project_resources(
     Ok(PreparedProjectService::new(
         project_id.to_owned(),
         service_id.to_owned(),
-        credential,
+        Some(credential),
         environment,
         container_environment,
         Some(route),